gemini_dia1/pyproject.toml
[project]
name = "gemini-dia1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.3",
]
